move batch changes resources to their own file
