+++ vinyldns/resources.go
@@ -49,55 +49,3 @@ type ListFilter struct {
 	StartFrom  string
 	MaxItems   int
 }
-
-// BatchRecordChanges represents a list of record changes,
-// as returned by the list batch changes VinylDNS API endpoint.
-type BatchRecordChanges struct {
-	BatchChanges []RecordChange `json:"batchChanges,omitempty"`
-}
-
-// RecordChange represents an individual batch record change.
-type RecordChange struct {
-	ID               string     `json:"id,omitempty"`
-	Status           string     `json:"status,omitempty"`
-	ChangeType       string     `json:"changeType,omitempty"`
-	RecordName       string     `json:"recordName,omitempty"`
-	TTL              int        `json:"ttl,omitempty"`
-	Type             string     `json:"type,omitempty"`
-	ZoneName         string     `json:"zoneName,omitempty"`
-	InputName        string     `json:"inputName,omitempty"`
-	ZoneID           string     `json:"zoneId,omitempty"`
-	TotalChanges     int        `json:"totalChanges,omitempty"`
-	UserName         string     `json:"userName,omitempty"`
-	Comments         string     `json:"comments,omitempty"`
-	UserID           string     `json:"userId,omitempty"`
-	CreatedTimestamp string     `json:"createdTimestamp,omitempty"`
-	Record           RecordData `json:"data,omitempty"`
-	OwnerGroupID     string     `json:"ownerGroupId,omitempty"`
-}
-
-// BatchRecordChangeUpdateResponse is represents a batch record change create or update response
-type BatchRecordChangeUpdateResponse struct {
-	Comments     string         `json:"comments,omitempty"`
-	OwnerGroupID string         `json:"ownerGroupId,omitempty"`
-	Changes      []RecordChange `json:"changes,omitempty"`
-}
-
-// RecordData is represents a batch record change record data.
-type RecordData struct {
-	Address  string `json:"address,omitempty"`
-	CName    string `json:"cname,omitempty"`
-	PTRDName string `json:"ptrdname,omitempty"`
-}
-
-// BatchRecordChange represents a batch record change API response.
-type BatchRecordChange struct {
-	ID               string         `json:"id,omitempty"`
-	UserName         string         `json:"userName,omitempty"`
-	UserID           string         `json:"userId,omitempty"`
-	Status           string         `json:"status,omitempty"`
-	Comments         string         `json:"comments,omitempty"`
-	CreatedTimestamp string         `json:"createdTimestamp,omitempty"`
-	OwnerGroupID     string         `json:"ownerGroupId,omitempty"`
-	Changes          []RecordChange `json:"changes,omitempty"`
-}
